more code refactoring for maintainabillity
