+++ .editorconfig
@@ -5,6 +5,8 @@ root = true
 charset = utf-8
 end_of_line = lf
 insert_final_newline = true
-indent_style = tab
+indent_style = space
 indent_size = 2
 tab_width = 2
+
+
