+++ .eslintrc
@@ -5,7 +5,8 @@
     "node": true
   },
   "plugins": [
-    "@typescript-eslint"
+    "@typescript-eslint",
+    "simple-import-sort"
   ],
   "extends": [
     "eslint:recommended",
@@ -17,6 +18,8 @@
   },
   "rules": {
     "no-unused-vars": 1,
+    "simple-import-sort/imports": "error",
+    "simple-import-sort/exports": "error",
     "@typescript-eslint/no-unused-vars": [
       "error",
       {
