+++ package.json
@@ -8,7 +8,11 @@
 		"build": "tsc -noEmit -skipLibCheck && node esbuild.config.mjs production",
 		"version": "node version-bump.mjs && git add manifest.json versions.json"
 	},
-	"keywords": ["obsidian", "contacts", "vcard"],
+	"keywords": [
+		"obsidian",
+		"contacts",
+		"vcard"
+	],
 	"author": "",
 	"license": "MIT",
 	"devDependencies": {
@@ -20,6 +24,8 @@
 		"@typescript-eslint/parser": "5.29.0",
 		"builtin-modules": "3.3.0",
 		"esbuild": "0.14.47",
+		"eslint-plugin-import": "^2.31.0",
+		"eslint-plugin-simple-import-sort": "^12.1.1",
 		"obsidian": "latest",
 		"tslib": "2.4.0",
 		"typescript": "4.7.4"
