+++ src/contacts/index.ts
@@ -1,0 +1,3 @@
+export * from './contactDataKeys';
+export * from './contactFrontmatter';
+export * from './contactMdTemplate';
