+++ src/contacts/vcard/vcardDefinitions.ts
