+++ src/context/sharedAppContext.ts
@@ -6,7 +6,6 @@ export function setApp(app: App) {
 	_app = app
 }
 
-// Safe accessor, runtime-enforced, TS knows it's not undefined
 export function getApp(): App {
 	if (!_app) {
 		throw new Error('App context has not been set.')
