--- src/parse/contact.ts
@@ -1,6 +1,0 @@
-import { TFile } from "obsidian";
-
-export type Contact = {
-	data: Record<string, any>;
-  file: TFile;
-}
