+++ src/ui/sidebar/components/ContactsListView.tsx
@@ -1,10 +1,10 @@
 import { randomUUID } from "crypto";
+import { TFile } from "obsidian";
 import * as React from "react";
-import { Contact } from "src/parse/contact";
+import { Contact } from "src/contacts";
+import { ContactView } from "src/ui/sidebar/components/ContactView";
 import { Sort } from "src/util/constants";
-import { ContactView } from "./ContactView";
 import myScrollTo from "src/util/myScrollTo";
-import {TFile} from "obsidian";
 
 
 type ContactsListProps = {
